Fusion

+++ src/uti_phgrm/SaisiePts/cAppliInput.cpp
@@ -90,7 +90,7 @@ void cX11_Interface::SetInvisRef(bool aVal)
 {
     mRefInvis = aVal;
 
-    for (uint aKW=0 ; aKW < mWins.size(); aKW++)
+    for (int aKW=0 ; aKW < (int)mWins.size(); aKW++)
     {
         mWins[aKW]->BCaseVR()->SetVal(aVal);
         mWins[aKW]->Redraw();
@@ -103,7 +103,7 @@ void cQT_Interface::SetInvisRef(bool aVal)
     mRefInvis = aVal;
 
     //TODO:
-    /* for (uint aKW=0 ; aKW < mWins.size(); aKW++)
+    /* for (int aKW=0 ; aKW < (int)mWins.size(); aKW++)
     {
         mWins[aKW]->BCaseVR()->SetVal(aVal);
         mWins[aKW]->Redraw();
@@ -142,7 +142,7 @@ cCaseNamePoint *cQT_Interface::GetIndexNamePoint()
 
 void cX11_Interface::RedrawAllWindows()
 {
-    for (uint aK=0 ; aK< mWins.size() ; aK++)
+    for (int aK=0 ; aK< (int)mWins.size() ; aK++)
         mWins[aK]->Redraw();
 }
 
@@ -190,7 +190,7 @@ bool cAppli_SaisiePts::Visible(cSP_PointeImage & aPIm)
 
 void cAppli_SaisiePts::HighLightSom(cSP_PointGlob * aPG)
 {
-    for (uint aKP=0 ; aKP< mPG.size() ; aKP++)
+    for (int aKP=0 ; aKP< int(mPG.size()) ; aKP++)
     {
         if (mPG[aKP] == aPG)
             aPG->HighLighted() = ! aPG->HighLighted();
@@ -201,7 +201,7 @@ void cAppli_SaisiePts::HighLightSom(cSP_PointGlob * aPG)
 
 void cAppli_SaisiePts::ChangeName(std::string anOldName, std::string  aNewName)
 {
-    for (uint aKP=0 ; aKP< mPG.size() ; aKP++)
+    for (int aKP=0 ; aKP< int(mPG.size()) ; aKP++)
     {
         if (mPG[aKP]->PG()->Name() == aNewName)
         {
@@ -210,7 +210,7 @@ void cAppli_SaisiePts::ChangeName(std::string anOldName, std::string  aNewName)
         }
     }
 
-    for (uint aKP=0 ; aKP< mPG.size() ; aKP++)
+    for (int aKP=0 ; aKP< int(mPG.size()) ; aKP++)
     {
         if (mPG[aKP]->PG()->Name() == anOldName)
         {
